# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MegaConvert API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /workflows/{id}/run:
    post:
      summary: Run workflow
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Started }
...
  /apikeys/{id}:
    delete:
      summary: Revoke API key
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204": { description: Revoked }
...